+++ src/main/scala/edu/ie3/simona/service/ServiceStateData.scala
@@ -6,10 +6,6 @@
 
 package edu.ie3.simona.service
 
-import edu.ie3.simona.ontology.messages.{Activation, SchedulerMessage}
-import edu.ie3.util.scala.collection.immutable.SortedDistinctSeq
-import org.apache.pekko.actor.typed.ActorRef
-
 trait ServiceStateData
 
 object ServiceStateData {
@@ -19,10 +15,4 @@ object ServiceStateData {
   trait InitializeServiceStateData extends ServiceStateData
 
   trait ServiceBaseStateData extends ServiceStateData
-
-  final case class ServiceConstantStateData(
-      scheduler: ActorRef[SchedulerMessage],
-      activationAdapter: ActorRef[Activation],
-  ) extends ServiceStateData
-
 }